tweak(layout): Split components and establish MainLayout
