+++ components/layout/Header.tsx
@@ -1,12 +1,8 @@
-import Head from 'next/head';
-import Link from 'next/link';
 import { ArrowRightIcon, HeartIcon, XIcon } from '@heroicons/react/outline';
 import React, { Fragment, useRef, useState } from 'react';
 import { Dialog, Transition } from '@headlessui/react';
 import CountUp, { CountUpProps } from 'react-countup';
-import { MenuAlt4Icon } from '@heroicons/react/outline';
-import { NavbarItemData, NavbarItems } from '../../static/config';
-
+import { MobileNavBar, DesktopNavBar } from './NavBar';
 interface StatCounterProps extends CountUpProps {
   labelSuffix?: string;
   valueSuffix: string;
@@ -44,41 +40,19 @@ const StatCounter: React.FC<StatCounterProps> = ({
   );
 };
 
-const MobileNavItem: React.FC<NavbarItemData> = ({ name, url }) => (
-  <a
-    className='text-gray-300 font-bold bg-gray-800 hover:bg-gray-700 transition duration-200 ease-in-out hover:text-white px-3 py-2 rounded-md text-sm font-medium'
-    href={url}
-    target='_blank'
-    rel='noopener'
-  >
-    {name}
-  </a>
-);
-
-const NavItem: React.FC<NavbarItemData> = ({ name, url }) => (
-  <a
-    className='duration-200 ease-in-out hover:bg-indigo-900 p-2 rounded-lg'
-    href={url}
-    target='_blank'
-    rel='noopener'
-  >
-    {name}
-  </a>
-);
-
 export default function Home() {
-  const [open, setOpen] = useState(false);
-  const [menuOpen, setmenuOpen] = useState(false);
+  const [zapDialogOpen, setZapDialogOpen] = useState(false);
+  const [mobileNavOpen, setMobileNavOpen] = useState(false);
   const cancelButtonRef = useRef(null);
 
   return (
     <div>
-      <Transition.Root show={open} as={Fragment}>
+      <Transition.Root show={zapDialogOpen} as={Fragment}>
         <Dialog
           as='div'
           className='fixed z-10 inset-0 overflow-y-auto'
           initialFocus={cancelButtonRef}
-          onClose={setOpen}
+          onClose={setZapDialogOpen}
         >
           <div className='flex items-center justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0'>
             <Transition.Child
@@ -118,7 +92,7 @@ export default function Home() {
                       <button
                         type='button'
                         className='hidden sm:block bg-gray-700 hover:bg-gray-600 transition duration-200 ease-in-out rounded-lg p-2'
-                        onClick={() => setOpen(false)}
+                        onClick={() => setZapDialogOpen(false)}
                         ref={cancelButtonRef}
                       >
                         <XIcon className='h-5 w-5 text-gray-300' />
@@ -179,7 +153,7 @@ export default function Home() {
                   <button
                     type='button'
                     className='sm:hidden mt-3 w-full inline-flex justify-center rounded-lg shadow-sm px-4 py-2 transition duration-200 ease-in-out bg-gray-700 text-base font-medium text-gray-200 hover:bg-gray-800 focus:outline-none focus:ring-0 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm'
-                    onClick={() => setOpen(false)}
+                    onClick={() => setZapDialogOpen(false)}
                     ref={cancelButtonRef}
                   >
                     Go Back
@@ -190,124 +164,8 @@ export default function Home() {
           </div>
         </Dialog>
       </Transition.Root>
-      <Transition.Root show={menuOpen} as={Fragment}>
-        <Dialog as='div' className='fixed inset-0 overflow-hidden' onClose={setmenuOpen}>
-          <div className='absolute inset-0 overflow-hidden'>
-            <Transition.Child
-              as={Fragment}
-              enter='ease-in-out duration-500'
-              enterFrom='opacity-0'
-              enterTo='opacity-100'
-              leave='ease-in-out duration-500'
-              leaveFrom='opacity-100'
-              leaveTo='opacity-0'
-            >
-              <Dialog.Overlay className='absolute inset-0 bg-gray-900 bg-opacity-75 transition-opacity' />
-            </Transition.Child>
-            <div className='fixed inset-y-0 right-0 pl-10 max-w-xs sm:max-w-sm flex'>
-              <Transition.Child
-                as={Fragment}
-                enter='transform transition ease-in-out duration-500 sm:duration-700'
-                enterFrom='translate-x-full'
-                enterTo='translate-x-0'
-                leave='transform transition ease-in-out duration-500 sm:duration-700'
-                leaveFrom='translate-x-0'
-                leaveTo='translate-x-full'
-              >
-                <div className='relative w-screen max-w-md'>
-                  <Transition.Child
-                    as={Fragment}
-                    enter='ease-in-out duration-500'
-                    enterFrom='opacity-0'
-                    enterTo='opacity-100'
-                    leave='ease-in-out duration-500'
-                    leaveFrom='opacity-100'
-                    leaveTo='opacity-0'
-                  >
-                    <div className='absolute top-0 left-0 -ml-8 pt-4 pr-2 flex sm:-ml-10 sm:pr-4'>
-                      <button
-                        type='button'
-                        className='rounded-md text-gray-300 duration-200 ease-in-out hover:text-white focus:outline-none focus:ring-0'
-                        onClick={() => setmenuOpen(false)}
-                      >
-                        <span className='sr-only'>Close panel</span>
-                        <XIcon className='h-6 w-6' aria-hidden='true' />
-                      </button>
-                    </div>
-                  </Transition.Child>
-                  <div className='h-full flex flex-col py-6 bg-gray-900 shadow-xl overflow-y-scroll'>
-                    <div className='px-4 sm:px-6'>
-                      <Dialog.Title className='text-lg font-medium text-gray-200'>
-                        Navigation
-                      </Dialog.Title>
-                    </div>
-                    <div className='mt-6 relative flex-1 px-4 sm:px-6'>
-                      <nav className='flex flex-col space-y-2'>
-                        {NavbarItems.map((item) => (
-                          <MobileNavItem
-                            url={item.url}
-                            name={item.name}
-                            key={item.name}
-                          />
-                        ))}
-                      </nav>
-                    </div>
-                  </div>
-                </div>
-              </Transition.Child>
-            </div>
-          </div>
-        </Dialog>
-      </Transition.Root>
-      <Head>
-        <title>txAdmin - The best FiveM server management solution</title>
-        <link rel='icon' href='./favicon.png' />
-        <meta
-          name='viewport'
-          content='width=device-width, initial-scale=1, shrink-to-fit=no'
-        />
-        <meta name='theme-color' content='#0EB880' />
-        <meta
-          name='description'
-          content='txAdmin is a completely free to use, full-featured web panel to Manage & Monitor your FiveM Server remotely, in use by over TEN thousand servers worldwide at any given time.'
-        />
-        <meta
-          name='keywords'
-          content='txadmin, fivem, fivem server, login, discord, manage fivem, fivem monitoring, fivem server panel, fivem web panel, fivem recipes'
-        />
-        <meta property='og:type' content='website' />
-        <meta property='og:site_name' content='txAdmin' />
-        <meta
-          property='og:title'
-          content='txAdmin - The best FiveM server management solution'
-        />
-        <meta
-          property='og:description'
-          content='txAdmin is a completely free to use, full-featured web panel to Manage & Monitor your FiveM Server remotely, in use by over TEN thousand servers worldwide at any given time.'
-        />
-        <meta property='og:url' content='https://txadm.in' />
-        <meta property='og:image' content='https://i.file.glass/yuFh9dmlBS.png' />
-        <link rel='canonical' href='https://txadm.in' />
-      </Head>
-      <div className='flex justify-between mx-5 md:mx-0 md:justify-around mt-5'>
-        <div>
-          <Link href='/' passHref>
-            <a className='font-bold text-lg select-none text-white'>txAdmin</a>
-          </Link>
-        </div>
-        <button
-          aria-label='Mobile Navigation Button'
-          className='flex md:hidden bg-gray-300 hover:bg-gray-200 duration-200 ease-in-out p-2 rounded-xl'
-          onClick={() => setmenuOpen(true)}
-        >
-          <MenuAlt4Icon className='h-6 w-6 text-gray-900' />
-        </button>
-        <nav className='hidden md:block space-x-1 text-base text-white font-medium'>
-          {NavbarItems.map((item) => (
-            <NavItem url={item.url} name={item.name} key={item.name} />
-          ))}
-        </nav>
-      </div>
+      <MobileNavBar isOpen={mobileNavOpen} onClose={() => setMobileNavOpen(false)} />
+      <DesktopNavBar handleOpenMobile={() => setMobileNavOpen(true)} />
       <div className='flex h-screen items-center justify-center'>
         <div className='m-auto max-w-3xl p-4'>
           <div className='flex flex-col space-y-4 align-middle text-center'>
@@ -349,7 +207,7 @@ export default function Home() {
                 </div>
                 <div>
                   <button
-                    onClick={() => setOpen(true)}
+                    onClick={() => setZapDialogOpen(true)}
                     className='group flex mx-auto justify-center w-60 bg-txgreen hover:bg-green-400 text-black font-bold text-base py-2 px-3 rounded-lg transition duration-200 ease-in-out'
                   >
                     Get Started
