+++ pages/_app.tsx
@@ -1,7 +1,12 @@
-import 'tailwindcss/tailwind.css'
-import type { AppProps } from 'next/app'
+import 'tailwindcss/tailwind.css';
+import type { AppProps } from 'next/app';
+import { MainLayout } from '../components/layout/MainLayout';
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  return (
+    <MainLayout>
+      <Component {...pageProps} />
+    </MainLayout>
+  );
 }
-export default MyApp
+export default MyApp;
